pyproject.toml
[build-system]
requires = ["hatchling>=1.25", "hatch-vcs>=0.4"]
build-backend = "hatchling.build"

[project]
name = "sabr-kit"
dynamic = ["version"]
description = "Structure-based Antibody Renumbering"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"

authors = [
  { name = "Diego del Alamo" },
]

dependencies = [
  "biopython>=1.85",
  "click>=8.1",
  "gemmi>=0.7",
  "jax>=0.4.20",
  "jaxlib>=0.4.20",
  "numpy>=2.4",
  "pandas>=2.2",
  "requests>=2.28",
  "scipy>=1.12",
  "dm-haiku>=0.0.12"
]

[project.optional-dependencies]
test = [
  "pytest>=7.0",
  "pytest-cov>=4.1",
]

[project.urls]
Homepage = "https://github.com/delalamo/sabr"
Issues = "https://github.com/delalamo/sabr/issues"

[project.scripts]
sabr = "sabr.cli:main"

# ----------------------------------------
# Include external submodules in the build
# ----------------------------------------
[tool.hatch.build.targets.wheel]
packages = ["src/sabr"]

[tool.hatch.build.targets.wheel.force-include]
"external/ANARCI/src/ANARCI" = "ANARCI"

# Optionally, include any data files you want in the wheel
[tool.hatch.build]
include = [
  "src/sabr/**",
  "external/ANARCI/src/ANARCI/**",
  "README.md",
  "LICENSE"
]

# ----------------------------------------
# Optional: version automatically from Git tags
# ----------------------------------------
[tool.hatch.version]
source = "vcs"

[tool.black]
line-length = 80
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 80

[tool.flake8]
max-line-length = 80
extend-ignore = ["E203", "W503"]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (skipped by default, use -m '' to include)",
]
addopts = "-m 'not slow'"
